move files to runner directory
